+++ Moyoy-Api/src/main/java/com/moyoy/api/common/filter/RateLimitingFilter.java
@@ -2,6 +2,19 @@ package com.moyoy.api.common.filter;
 
 import static com.moyoy.common.constant.MoyoConstants.*;
 
+import java.io.IOException;
+import java.util.Optional;
+import java.util.concurrent.TimeUnit;
+
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Component;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import com.moyoy.api.common.util.ErrorResponseWriter;
+
+import com.moyoy.domain.support.error.CommonErrorCode;
+
 import io.github.bucket4j.Bucket;
 import io.github.bucket4j.BucketConfiguration;
 import io.github.bucket4j.ConsumptionProbe;
@@ -10,17 +23,6 @@ import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
-import java.util.Optional;
-import java.util.concurrent.TimeUnit;
-
-import com.moyoy.api.common.util.ErrorResponseWriter;
-import com.moyoy.domain.support.error.CommonErrorCode;
 
 @Component
 public class RateLimitingFilter extends OncePerRequestFilter {
@@ -48,8 +50,7 @@ public class RateLimitingFilter extends OncePerRequestFilter {
 
 			response.setHeader("X-Rate-Limit-Remaining", String.valueOf(probe.getRemainingTokens()));
 			filterChain.doFilter(request, response);
-		}
-		else {
+		} else {
 
 			long waitNanos = probe.getNanosToWaitForRefill();
 			long retryAfterSeconds = TimeUnit.NANOSECONDS.toSeconds(waitNanos) + 1;
