+++ Moyoy-Api/src/main/java/com/moyoy/api/github_follow/application/response/GithubFollowDetectionResult.java
@@ -2,7 +2,7 @@ package com.moyoy.api.github_follow.application.response;
 
 import java.time.LocalDateTime;
 
-import com.moyoy.domain.follow.GithubUser;
+import com.moyoy.domain.github_follow.GithubUser;
 import com.moyoy.domain.support.page.SliceResult;
 
 public record GithubFollowDetectionResult(
