+++ Moyoy-Batch/src/main/java/com/moyoy/batch/dto/UserRankResult.java
@@ -1,6 +1,6 @@
 package com.moyoy.batch.dto;
 
-import com.moyoy.domain.ranking.RankingCalculatorResult;
+import com.moyoy.domain.ranking.dto.RankingCalculatorResult;
 
 public record UserRankResult(
 	Long userId,
