+++ Moyoy-Batch/src/main/java/com/moyoy/batch/job/RankCalculationJobConfig.java
@@ -40,8 +40,8 @@ import com.moyoy.batch.helper.GithubCommitStatCalculator;
 import com.moyoy.batch.listener.RankingJobDiscordListener;
 import com.moyoy.domain.ranking.GithubCommitStats;
 import com.moyoy.domain.ranking.RankingCalculator;
-import com.moyoy.domain.ranking.RankingCalculatorParameters;
-import com.moyoy.domain.ranking.RankingCalculatorResult;
+import com.moyoy.domain.ranking.dto.RankingCalculatorParameters;
+import com.moyoy.domain.ranking.dto.RankingCalculatorResult;
 import com.moyoy.infra.database.mysql.user.UserEntity;
 import com.moyoy.infra.external.github.helper.GithubApiLimitChecker;
 import com.moyoy.infra.external.github.helper.GithubOAuthTokenReader;
