+++ Moyoy-Domain/src/main/java/com/moyoy/domain/github_follow/GithubFollowClassifier.java
@@ -1,4 +1,4 @@
-package com.moyoy.domain.follow;
+package com.moyoy.domain.github_follow;
 
 import java.util.ArrayList;
 import java.util.List;
