+++ Moyoy-Domain/src/main/java/com/moyoy/domain/github_follow/GithubUser.java
@@ -1,4 +1,4 @@
-package com.moyoy.domain.follow;
+package com.moyoy.domain.github_follow;
 
 public record GithubUser(
 	Integer id,
