+++ Moyoy-Domain/src/main/java/com/moyoy/domain/ranking/Ranking.java
@@ -1,5 +1,7 @@
 package com.moyoy.domain.ranking;
 
+import com.moyoy.domain.ranking.dto.RankingUpdate;
+
 import lombok.Builder;
 import lombok.Getter;
