+++ Moyoy-Domain/src/main/java/com/moyoy/domain/ranking/dto/RankingCalculatorResult.java
@@ -1,4 +1,4 @@
-package com.moyoy.domain.ranking;
+package com.moyoy.domain.ranking.dto;
 
 public record RankingCalculatorResult(
 	long weekRankingPoint,
