+++ Moyoy-Infra/src/main/java/com/moyoy/infra/database/mysql/ranking/RankingRepositoryImpl.java
@@ -24,6 +24,20 @@ public class RankingRepositoryImpl implements RankingRepository {
 	private final RankingQueryDslRepository rankingQueryDslRepository;
 
 	@Override
+	public Optional<Ranking> findById(Long id) {
+
+		Optional<RankingEntity> rankingEntity = rankingJpaRepository.findById(id);
+		return rankingEntity.map(RankingMapper::toModel);
+	}
+
+	@Override
+	public Optional<Ranking> findByUserId(Long userId) {
+
+		RankingEntity rankingEntity = rankingJpaRepository.findByUserId(userId);
+		return Optional.ofNullable(RankingMapper.toModel(rankingEntity));
+	}
+
+	@Override
 	public void save(Ranking ranking) {
 
 		RankingEntity rankingEntity = RankingMapper.toEntity(ranking);
@@ -37,12 +51,7 @@ public class RankingRepositoryImpl implements RankingRepository {
 		rankingJpaRepository.saveAll(rankingEntityList);
 	}
 
-	@Override
-	public Optional<Ranking> findById(Long userId) {
 
-		RankingEntity rankingEntity = rankingJpaRepository.findByUserId(userId);
-		return Optional.ofNullable(RankingMapper.toModel(rankingEntity));
-	}
 
 	@Override
 	public SliceResult<Ranking> findAll(RankingPeriod duration, PageData pageData) {
