+++ Moyoy-Infra/src/main/java/com/moyoy/infra/database/redis/support/config/CacheConfig.java
@@ -17,7 +17,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 
-import com.moyoy.domain.follow.GithubFollowSnapshot;
+import com.moyoy.domain.github_follow.GithubFollowSnapshot;
 
 /**
  *   [추후 확장시 고려 사항]
